pyproject.toml
[project]
name = "ai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "graphviz>=0.20.3",
    "matplotlib>=3.10.1",
    "neat-python>=0.92",
    "pygame>=2.6.1",
    "shutils>=0.1.0",
]